Simplify core modules and remove unused legacy code
