+++ app_legacy/exchange_adapter/base.py
